Also properly perform backfill for coinbase events
While technically coinbase events were filled in before, those were only
filled in for blocks with missing transactions. We need an extra set of
steps to make sure coinbase transactions get backfilled as well.

Also, we include a field in the commandline options for backfill called
`eventsActivationHeight` because events can be only be seen after a
certain block height. If the field is not provided, we default to the
first block height events (for coinbase, in particular) can be found in
for the mainnet network.