add unsafeEventsQueryStmt
It is unsafe because we're not performing a join with the blocks table.
This join was "necessary" to ensure that queried event was matched to a
known block (even if it is an orphan). There is an extra query that can
be dispatched after the fact to ensure the queried event is associated
with a known block (it is called `getBlockTimes`).