Adding accounts transfer history query endpoint v2 (#76)
* point to appropriate commit in chainweb-api

* include accountHandler function

* fix cabal.project

* create Transfer table

* add module name to row in Transfer table

* write function mkTransferRows

worker function for writing rows to transfers table

* write transfers to db

* checking the params field appopriately

It is ok if there are only three values in this field (given the other
conditions of course)

* hardcode the events activation height map

this is necessary because the actual query takes some non-trivial amount
of time to execute and we need this map during backfills, listens, etc.
These are processes which should not be unnecessarily sluggish.

* introduce AccountDetail type for account queries

there is a build error. I need to figure out how to make a function with
this type "PrimaryKey BlockT Identity -> Text". using unDbHash and
coerce doesn't seem to work.

* fill hole in _acDetail_blockHash

* update cabal.project

* convert int pact object into a double

* include FillTransfers.hs module

we need this in case there are gaps in the new upcoming transfers table

* etch out basics for only filling transfers from an empty state

* first draft of filltransfers

filling an empty table

* fix mkTransferRows to include coinbase events

* update writePayload function to account for events

* point to chainweb-api master

* only pass in version instead of entire env to insertNewHeader

* makeEventsMinHeightMap -> eventsMinHeight

* use eventsMinHeight in fillEventsCut

* _tr_name -> _tr_modulename

* print raw value in mkTransferRows on error

* put in better debugging info

* index into json array properly

* take out fromJust uses

* Maybe UTCTime -> Either String UTCTime

* funnel creationTime for coinbase transactions properly

* make _tr_block field correct name in postgres

* "decimal" value can be a number rendered as a string

We can encounter this case

{ "decimal" : String "42.02472052" }

For this, we have correctly detect the String case and act accordingly

* use decimal/int codecs from chainweb-api

* don't proceed w/ fill if transfers table exists

the other filling capabilities should suffice if there are gaps in the
transfers table (fill on events)

* backfill transfers

* fix spelling error

* add cmdline option for backfill-transfers

Also make sure the backfill uses at the minimum available height for
every chain instead of the maximum.

* join transactions and events tables along requestkey

* correct trantfer insertion process

* update cabal.project

* parallelize along chunks instead of chainid

Through some experimentation, we deem it unnecessary to group transfer
table rows by chainid. Also, it is faster to not do a join along the
events and transactions tables to get the correct info to insert for the
transfers table.

* change filling strategy for transfers

use limits and offsets to easily skip non-candidate rows from the
events table.

* disable indexes during transfers backfill

* add moduleHash to primary key

* label column correctly in primary key description

* make transfer rows more unique

* fill in smaller chunks on individual chains

* correct filling strategy

calculate limits and offsets correctly. We also have to change how we
schedule work.

* correct filling strategy pt. 2

* fill on all chains  for now

* integrate doug's suggestions pt.1

* Max -> Min

* simplify code

* now we give the user an overestimate of how rows we need to fill, so
that we never give them a false reading of how much time it should take
to finish the transfer fill (`progress` would give negative times
before!)

* simple clean ups here and there

* optimize transfer insertions

* get argument order correct

* delete FillTransfers.hs

* cleanup work after deleting FillTransfers.hs

* optimize account search query and create indexes for it

* use withEventsMinHeight utility to simplify code

* improve indexes for account search query

* optimize accountSearch Query (pt2)

* update README

* cleanup mkTransferRows

* fix amount bug

* change amount type from Double to KDAScientific

This gets us the right amount of numeric precision needed!

* make sure to check if the length of params is 3

* update chainweb-api types

* merge master from chainweb-api

amount type :: Double -> Scientific

* Change numeric type of transfer amount type to Decimal (#94)

* update chainweb-api types

* Change numeric type of transfer amount type to Decimal

* move back to using Scientific

* specify correct HasColumnType instance for KDAScientific

* show hashes in errors of payloadWithOutputsBatch

* show hash properly

* bring in new commit from cw-api to fix chainid issue

* include ability to run "unsafe" migration

In this instance, I want to be able to run a migration that doesn't give
any warnings about user made edits to the chainweb-data db schema.

* improve logging in writeBlocks function

* bump chainweb-api pin

* bugfix: create non-coinbase transfers correctly

* some cleanup

* update endpoint definition (from chainweb-api)

* swap arguments

* log account search origination

* get printf correct

* include fromHeight logic

* print offset error message

* update cabal.project file

* point to master branch on chainweb-api

* update cabal.project again

Co-authored-by: Doug Beardsley <mightybyte@gmail.com>