Extend the account search endpoint with continuation (#103)
This PR extends the `/txs/account` endpoint to allow a simple, robust and efficient way to fetch the results of an account transfers search in a sequence of batches. Whenever a `/txs/accounts` request produces more results than the `limit` parameter, the response contains a `Chainweb-Next` response header containing a Base64-URL-encoded token. This token can then be passed to subsequent search requests to the same account via the `next` query parameter in order to fetch any further results.

Unlike the `limit` & `offset` way of fetching result batches, this method is more efficient, because the encoded token contains hints used for making it efficient to resume the search and it's also more robust for the client, because even if new blocks appear on the network, subsequent fetches should return a mutually exclusive set of rows, avoiding potential visual glitches or duplicated data.

* Extend accountQueryStmt with continuation

* Generalize tupleCmp

* Plumb the next tokens through the account handler

* Use reqkey for unambiguous ordering of transfers

* base64 URL encode the token

* Encode the whole token

* Factor out the token encoding utils

* Update the chainweb-api pin to point to master