Avoid error caused by the schema_migrations table
In preparation for switching to the script-based migration, we've added
the creation of the schema_migrations table to the cw-data start up.
However, when we add that table, the next time we start chainweb-data
it complains about the need to remove that table, which is obviously
not something we intend to do.

This PR filters the calculated migration steps and removes any
"TableRemove" step. Note that this should avoid the spurious
"DROP schema_migrations" error when the database is otherwise fully
migrated, but if there are any additional migrations required, cw-data
will still try to perform those migrations using
tryRunMigrationsWithEditUpdate, which will in turn error out due to the
spurious TableRemove steps. However, since this is supposed to be the
final version of cw-data that will perform migrations with
beam-automigrate, our users shouldn't ever run into that problem.