Implement chunked processing for event and tx search endpoints (#109)
This PR extends the `/txs/search` and `/txs/events` endpoints such that when the request contains the header `Chainweb-Execution-Strategy: Bounded`, the SQL queries performed by these endpoints can only scan a bounded number of rows from the underlying database tables. When the header is missing or has the value `Unbounded`, the behavior is the same as before.

CW-D instances accessible by untrusted users can currently overload the database easily by searching for something that doesn't exist or only exists very early in the blockchain history. The reason is that these queries scan through the source tables looking for matching rows and they will keep scanning until enough rows are found. 

The intention here is that a CW-D operator can choose to expose an instance directly and the users can call these endpoints as usual and the behavior (and the potentially poor performance) would be backward compatible. But the operator also has the choice to expose the instance behind an API gateway that always adds the `Chainweb-Execution-Strategy: Bounded` header so that any given API request has an upper bound on the amount of load it can put on the database. This allows the API gateway to employ rate limiting with predictable and fair resource allocation.

When `Chainweb-Execution-Strategy: Bounded` is passed, these search endpoints are allowed to return fewer rows than the request specifies. For example, a `limit=10` query might end up receiving fewer than 10 results even though the search matches more than 10 rows in the database. When that happens, CW-D responds with a `Chainweb-Next` header containing a token that can be passed into a subsequent query into the same endpoint through the `next` query parameter, making CW-D resume the search for further rows from the point where it left the last time.

Note that existing CW-D clients might be making assumptions about the case when they receive fewer rows than asked for with the `limit` parameter and they wouldn't know to look for the `Chainweb-Next` header, so this breaks backward compatibility for those clients; They need to be updated to make subsequent requests with the `next` parameter.

* Extend search endpoint types with Chainweb-Next

* Refactor evHandler for Chainweb-Next

* Implement bounded scan event search

* Minor type simplification

* Fix the chainweb-api pin sha256

* Factor out boundedScanOffset as a reusable utility

* Factor out boundedScanLimit as a reusable utility

* Move boundedScan utilities to new module

* Minor style improvement

* Refactor searchTxs in preperation for boundedScan

* Fix the empty event search error

I initially thought that it could only happen when the database is
empty, but it can also happen when the user makes a query that will
target zero rows, (such as minheight>current height)

* Fix minheight behaving like maxheight

* Factor out the bounded scan execution logic

* Minor style improvement

* Implement chunked processing for tx search

* Rename ev references to row in BoundedScan

That's a naming artifact from before the refactor

* Remove unnecessary where in boundedScanOffset

Seems like with more monomorphic types that's not necessary anymore

* Minor code layout improvement

* Remove unnecessary RankN arg in boundedScanOffset

* Generalize the types of boundedScan queries

We don't need the flexibility from this generalization yet

* Refactor BoundedScan to give it simpler semantics

* Improve variable names in BoundedScan

* Move code around in BoundedScan for presentation

* Rename P to Exp in BoundedScan

* Avoid BoundedScan dep on ChainwebData.Pagination

* Fix next + offset error message in /txs/search

* Minor style improvements

* Push full FromBackendRow constratint to call site

We want the entire FromBackendRow instance for the limit query to be
assembled at the call site of
performBoundedScan for more optimization opportunities, that
instance is used for parsing every single row after all

* Implement unbounded interpretation for BoundedScan

* Reorder module exports for presentation

* Add unbounded execution to performBoundedScan

* Add explanations to key definitions in BoundedScan

* Read the BoundedScan exec strategy from header

* Avoid fragile repetition in cursor definitions

* Push cursor continuation into performBoundedScan

* Refactor mkContinuation out of search,ev handlers

That logic is now purely about continuations

* Decompose BoundedScan and get rid of it

* Fix order direction bug

* Join blocks table only for matching events

* Remove accidentally added file

* Minor style improvements

* Move result decoration into the scan layer

This way, the OFFSET, LIMIT and ORDER BY clauses apply to the decorated
result, which should result in nicer behavior when decorations are not
1:1 and it also makes sure that the output order isn't altered due to a
new surrounding SELECT

* Fix comparison operator in cont calculation

* Update the chainweb-api pin

* Handle the offset=0 case properly