Refactor accountQueryStmt for bounded scans (#117)
This PR refactors `accountQueryStmt` to decompose it into the parts that define a bounded scan. We'll be pushing another PR to actually convert the accounts endpoint into a bounded scan, but this refactor makes sure that Postgres is able to produce an equivalent query plan for this new shape that's ready to be turned into a bounded scan.

Note that in addition to moving `WHERE` (`guard_`) conditions around, this refactor also removes the `limit_` clause on the individual branches of the `transfers` subquery. The reason for the existence of that inner `limit_` was that we had initially thought Postgres would end up fetching all the rows on both branches of the `UNION ALL` and then merge the rows from the two branches, so we wanted to put an upper bound on how many rows it would fetch from those branches. However, it turns out that Postgres is actually much smarter than that: It uses a `Merge Append` node in the query plan for implementing the `UNION ALL`  and `Merge Append` walks through the branches in lock step and fetches them and zips them based on the sort order simultaneously, so that means no inner `LIMIT` clause was necessary (and thankfully so, since that wouldn't be bounded-scan friendly!).