Add migrate command to chainweb-data CLI (#115)
This PR is a part of the incremental script-based migration transition #101.

This PR adds a `migrate` subcommand to `chainweb-data` that can be used to just run migrations and exit. This will be useful for the post-script-based-migration version; When that version encounters a database that's too old, it'll ask the user to download the latest `beam-automigrate`-based CW-D version (the upcoming release) and run it in this PR's `migrate` mode so that the database can be used with future versions that know how do migrate it incrementally from the last `beam-automigrate`-based database state.