Modify account search with bounded scan mechanics (#121)
This PR extends the `/txs/account` endpoint by applying the `BoundedScan` machinery to it. This way, when the `Chainweb-Execution-Strategy: Bounded` header is passed to it, it will only scan up to a predetermined number of rows in the database, just like the `/txs/events` and `/txs/search` endpoints.

Note that under normal circumstances, the bounded scan should be unnoticeable for the client of the `/txs/account` endpoint, because unlike the search endpoints (`/txs/{events,search}`), `/txs/account` fetches is results almost directly from the respective `account,height` indexes. However, it still applies some extra conditions that aren't directly on the indexes (such as the token type and the optional chain ID predicates). For the vast majority of the practical cases, these predicates shouldn't filter out enough rows to make the bounded scan query produce less rows than the `limit`, that's why we're currently not trying to address this concern by extending our indexes (and incurring other costs). So, having the bounded scan machinery under this endpoint makes sure that specifically crafted `/txs/account` queries can't hit various corner cases that make the server do a lot of work per request.

* modify account search with bounded mechanics

* Remove the obsoleted accountQueryStmt

* Whitespace

Co-authored-by: Enis Bayramoğlu <enisbayramoglu@gmail.com>