Harden the semantics of the bounded scan queries (#118)
This PR improves the SQL queries used by the bounded scan machinery to make them semantically unambiguous. Before this PR, as mentioned in [this stack overflow question](https://stackoverflow.com/questions/74800821/using-the-same-order-as-a-subquery-without-the-results-getting-sorted-unnecessar), we were running queries whose results theoretically depended on the query plan, so we were relying on Postgres not to go out of its way to break our query while it was semantically free to do so. Fortunately, this PR improves the queries to remove that ambiguity while maintaining the efficient query plans we had before. 

This PR achieves that by adding an `ORDER BY` clause to the outermost `SELECT`, disambiguating the rows that will get selected by the `LIMIT`, and avoids losing the efficient query plan (as mentioned in the stackoverflow question) by making sure that the inner query also has a `LIMIT` clause. 

I believe, that inner `LIMIT` clause hints to the query planner about the size of that subquery and it also makes sure that that the inner `ORDER BY` doesn't get ignored (due to the `LIMIT`).