Set random_page_cost to 0 for CW-D connections (#122)
Fixes #123

This PR initializes every new connection created by CW-D by executing `SET random_page_cost = 1.0`. This tells postgres to assume that accesses to random disk pages is equally expensive as accessing sequential pages. This is generally a good setting for a database that's storing its data on an SSD, but our intention here is to encourage Postgres to use index scan over table scans. The justification is that we design CW-D indexes and queries in tandem carefully to make sure all requests are serviced with predictable performance characteristics.