Gather XChain info for transfers (#137)
This PR extends the transfer search queries to attach cross-chain information to cross-chain transfer rows. For the send-side of cross-chain transactions it extracts the intended receiver from the corresponding `TRANSFER_XCHAIN` event of the same transaction and for the receive-side, it chases the `pactid` of the associated transaction and looks for a matching `TRANSFER_XCHAIN`. 

* Pin xchain info chainweb-api commit

* Join cross-chain toAcct to transfers

* Improve the gather subquery

* Implement xChainInfo for both directions

* Fix typo in comments

* Rename xchain -> crossChain