Separate ETL and HTTP services (#149)
In this PR, our goal is to make it possible for an operator to run chainweb-data either as a server or as an
ETL service gathering data from chainweb-node. We will still allow an operator to run chainweb-data performing the aforementioned services simultaneously.

* Separate ETL and HTTP services

In this PR, our goal is to make it possible for an operator
to run chainweb-data either as a server or as an
ETL service gathering data from chainweb-node. We will still allow
an operator to run chainweb-data performing the aforementioned services
simultaneously.

* clean up ServerEnv munging during executable startup

* make etl/http services separate commands

* rename Server|ServerEnv to Worker|WorkerEnv

* enis' suggestions part 1.

remove lenses & prisms.
make apiServer depend solely on the HTTPEnv datatype.

* remove unnecessary logging stmt

* simplify commands so that the etl/http stages can
be run under listen/server. There is a "no-etl" flag under
the server command to run http. Also, the user can provide
arguments to the listen command to affect how the etl is run
under listen.

* remove httpP

* remove export line

* Update haskell-src/lib/ChainwebData/Env.hs

Co-authored-by: Enis Bayramoğlu <enisbayramoglu@gmail.com>

---------

Co-authored-by: Enis Bayramoğlu <enisbayramoglu@gmail.com>