Work around HLS bug while embedding files
This commit works around the following HLS bug:

https://github.com/haskell/haskell-language-server/issues/481

That issue is actually related to stack, but seems like the cabal
integration is suffering from the same in our case. There are quite a
few relative-path related issue open HLS anyway.

This commit uses the `makeRelativeToProject` function from file-embed
in order to resolve the relative paths in a way that's independent of
the build system, working around this long-standing HLS bug.