HLS integration fixes for the chainweb-data codebase (#157)
This PR adds a `hie.yaml` to the project and also works around the following HLS bug:

https://github.com/haskell/haskell-language-server/issues/481

That issue is actually related to stack, but seems like the cabal integration is suffering from the same in our case. There are quite a few relative-path related issues open in HLS anyway.

This commit uses the `makeRelativeToProject` function from file-embed ([As suggested by phadej under another issue](https://github.com/commercialhaskell/stack/issues/5421#issuecomment-737110569)) in order to resolve the relative paths in a way that's independent of the build system, working around this long-standing HLS bug.

* Add a hie.yaml to the project

* Work around HLS bug while embedding files

This commit works around the following HLS bug:

https://github.com/haskell/haskell-language-server/issues/481

That issue is actually related to stack, but seems like the cabal
integration is suffering from the same in our case. There are quite a
few relative-path related issue open HLS anyway.

This commit uses the `makeRelativeToProject` function from file-embed
in order to resolve the relative paths in a way that's independent of
the build system, working around this long-standing HLS bug.